refactor(client): flatten package structure
