--- packages/client/src/content/index.ts
@@ -1,1 +1,0 @@
-export * from './upload'
