--- packages/client/src/content/upload/index.ts
@@ -1,1 +1,0 @@
-export * from './GenerateUploadUrlMutationBuilder'
